# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cobranca-API — Offline (engine pyCobranca)
  description: |
    Superfície **offline** da plataforma **Cobranca-API** (serviço único **100% Python**):
    geração de boletos bancários brasileiros, arquivos CNAB e parsing de extratos OFX —
    servida **nativamente** pela engine [pyCobrança](https://github.com/Maxwbh/pyCobranca)
    dentro do próprio processo FastAPI (sem sidecar, sem proxy).

    > **Nota de versão (v2.0.0):** os contratos desta superfície (paths, campos,
    > headers `X-*`) permanecem compatíveis com a v1.x. A única mudança de escopo é
    > o formato de saída: a engine emite **PDF** — `jpg`/`png`/`tif` respondem
    > **400**.

    ## Offline (`/api/*`) × Gateway (raiz) — quando usar cada um

    | | **Offline (esta doc)** | **Gateway REST** ([Swagger `/docs`](/docs)) |
    |---|---|---|
    | O que faz | GERA o documento (boleto/CNAB) localmente | REGISTRA a cobrança na API online do banco (C6, Sicoob, Inter, Itaú) |
    | Boleto | Não registrado (o banco só conhece no pagamento) | **Registrado** (CIP, em tempo real) + Pix, Bolepix, webhooks |
    | Autenticação | **Nenhuma** (por design — não toca em banco nem segredo) | Token `bapi_` (`POST /credenciais`) |
    | Conciliação | Retorno CNAB / OFX (arquivos) | Webhooks + consultas online |

    > **Autenticação:** esta superfície não tem `securitySchemes` de propósito.
    > O mecanismo `bapi_` pertence ao gateway e está documentado no
    > [Swagger do gateway](/docs).

...

...

...

...

...
servers:
...
paths:
...
  /api/boleto/validate:
    get:
      tags:
        - Boleto
      summary: Validar dados do boleto
      description: |
        Valida os dados sem gerar o arquivo. Em erro devolve **400** com
        `validation_errors` — uma **lista**, um item por campo violado.
        Regras por banco (tipos/tamanhos/carteiras/especiais) em
        [validacao-campos.md](https://github.com/Maxwbh/cobranca-api/blob/main/docs/api/validacao-campos.md).
      operationId: validateBoleto
      parameters:
        - name: bank
          in: query
          required: true
          description: Código do banco
          schema:
            $ref: '#/components/schemas/BankCode'
        - name: data
          in: query
          required: true
          description: Dados do boleto em JSON (URL-encoded)
          schema:
            type: string
            format: json
          example: '{"valor":150.0,"cedente":"Empresa Teste LTDA","documento_cedente":"11222333000181","sacado":"João da Silva","sacado_documento":"52998224725","agencia":"3073","conta_corrente":"12345678","convenio":"1234567","carteira":"18","nosso_numero":"123","data_vencimento":"2027-12-30"}'
      responses:
        '200':
          description: Dados válidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationSuccess'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...
  /api/render/boleto:
    post:
      tags:
        - Render
      summary: Renderizar boleto (dados + PDF base64)
      description: |
        Registra/calcula os dados do boleto (linha digitável, código de barras,
        nosso número) e devolve o PDF em base64. Corpo JSON, renderizado
        in-process pela engine pyCobrança. Consumido pelo provider offline do gateway.
      operationId: renderBoleto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bank, data]
              properties:
                bank:
                  type: string
                  example: itau
                data:
                  type: object
                  description: Dados do boleto (mesmos campos de /api/boleto)
                template:
                  type: string
                  enum: [moderno, classico]
                  default: moderno
                  description: >-
                    Modelo visual. O padrão `moderno` traz o logo do banco emissor
                    no cabeçalho.
      responses:
        '200':
          description: Boleto renderizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  nosso_numero: { type: string, example: '00000001' }
                  linha_digitavel: { type: string }
                  codigo_barras: { type: string }
                  pdf_base64: { type: string, description: PDF em base64 }
        '400':
          description: Dados do boleto inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
...
  schemas:
    BankCode:
      type: string
      enum:
        - banco_brasil
        - itau
        - bradesco
        - caixa
        - santander
        - sicoob
        - sicredi
        - banrisul
        - banestes
        - banco_nordeste
        - banco_brasilia
        - banco_c6
        - unicred
        - credisis
        - safra
        - citibank
        - hsbc
        - ailos
      description: Código do banco suportado (18 bancos brasileiros)
...
    ValidationSuccess:
      type: object
      properties:
        valid:
          type: boolean
          example: true
        message:
          type: string
          example: Dados do boleto são válidos
    ValidationError:
      type: object
      properties:
        valid:
          type: boolean
          example: false
        validation_errors:
          type: array
          description: |
            Lista plana — **um item por campo violado** (contrato da engine
            PyCobrança ≥ 1.0.1, que expõe `BoletoInvalido.erros`). Mapeie cada
            string ao campo do formulário. Nunca é um objeto por chave.
          items:
            type: string
          example:
            - "carteira '999' não suportada (use uma de: 104, 109, 112, 115, 175, 177, 188)"
            - agência deve ter no máximo 4 dígitos
            - conta deve ter no máximo 5 dígitos
        hint:
          type: string
          example: Corrija os erros de validação antes de gerar o boleto
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Mensagem de erro
          example: Dados inválidos
        validation_errors:
          type: array
          description: Lista plana — um item por campo violado (contrato PyCobrança ≥ 1.0.1).
          items:
            type: string
        details:
          type: array
          description: |
            Só em `/api/retorno`, que usa este nome no lugar de
            `validation_errors` — um item por problema do arquivo.
          items:
            type: string
        hint:
          type: string
          description: O que fazer, quando a rota tem o que sugerir.
          example: Verifique se todos os campos obrigatórios estão preenchidos